+++ mlx9064x/benches/load_calibration.rs
@@ -1,7 +1,11 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 
-use mlx9064x::{mlx90640::Mlx90640Calibration, mlx90641::Mlx90641Calibration};
-use mlx9064x_test_data::{mlx90640_datasheet_eeprom, mlx90641_datasheet_eeprom};
+use mlx9064x::{
+    mlx90640::Mlx90640Calibration, mlx90641::Mlx90641Calibration, mlx90642::Mlx90642Calibration,
+};
+use mlx9064x_test_data::{
+    mlx90640_datasheet_eeprom, mlx90641_datasheet_eeprom, mlx90642_datasheet_eeprom,
+};
 
 pub fn criterion_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("Calibration Loading");
@@ -11,6 +15,9 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     });
     group.bench_with_input("MLX90641", &mlx90641_datasheet_eeprom(), |b, eeprom| {
         b.iter(|| Mlx90641Calibration::from_data(eeprom))
+    });
+    group.bench_with_input("MLX90642", &mlx90642_datasheet_eeprom(), |b, eeprom| {
+        b.iter(|| Mlx90642Calibration::from_data(eeprom))
     });
 }